==> Rmarkdown/1.0-jvelezmagic-eda_experiments.Rmd <==
---
title: "1.0 - Experiment Data Exploration"
author: "Jesús Vélez Santiago"
date: "`r format(Sys.Date(), '%Y-%m')`"
output: 
  html_document:
    theme: readable
    highlight: kate
    toc: true
    toc_float: true
    toc_depth: 3
    code_folding: hide
    self_contained: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = TRUE,
  message = FALSE,
  warning = FALSE,
  fig.align = "center",
  # dev = "svg",
  fig.retina = 2
)
```

## Libraries

```{r libraries, message=FALSE}
library(tidyverse)
library(ggpubr)
library(ggwaffle)
library(here)
library(glue)
```

## Load Data

### Processed lineages

```{r load_processed_lineages}
processed_lineages_file <- here("data", "processed", "lineages.tsv")
processed_lineages_df <- read_tsv(processed_lineages_file, show_col_types = FALSE) %>% 
  mutate(across(contains("id"), as.character)) %>% 
  glimpse()
```

### Processed tracks

```{r load_processed_track}
tracks_summary_file <- here("data", "processed", "tracks_summary.tsv")
tracks_summary_df <- read_tsv(tracks_summary_file, show_col_types = FALSE) %>% 
  mutate(
    across(contains("id"), as.character),
    across(contains("filamented_"), ~factor(
        .x,
        levels = c(FALSE, TRUE),
        labels = c("Not filamented", "Filamented")
      )
    ),
    across(contains("survived_"), ~factor(
        .x,
        levels = c(FALSE, TRUE),
        labels = c("Not survived", "Survived")
      )
    ),
    cell_status = paste0(filamented_track, " - ", survived_strict) %>% 
      factor()
  ) %>%
  select(
    experiment_id, id,
    cell_status, filamented_track,
    contains("time"), contains("survived")
  ) %>% 
  glimpse()
```

## Minimal preprocessing

```{r minimal_preprocessing}
processed_lineages_df_2 <- processed_lineages_df %>% 
  mutate(
    across(contains("filamented_"), ~factor(
        .x,
        levels = c(FALSE, TRUE),
        labels = c("Not filamented", "Filamented")
      )
    )
  ) %>% 
  inner_join(tracks_summary_df, by = c("experiment_id", "id", "filamented_track")) %>% 
  relocate(where(is.character), where(is.factor)) %>% 
  glimpse()
```

## Exploratory Data Analysis

### Set default plot style

```{r default_plot_theme}
theme_set(
  theme_bw() +
  theme(legend.position = "top")
)
```

### Status by time

```{r}
status_time_df <- processed_lineages_df_2 %>% 
  select(experiment_id, id, time) %>% 
  group_by(experiment_id) %>% 
  group_modify(~complete(expand(.x, id, time))) %>% 
  ungroup() %>% 
  left_join(
    y = processed_lineages_df_2 %>% 
      select(experiment_id, id, time, filamented_at_frame, initial_time, end_time),
    by = c("experiment_id", "id", "time")
  ) %>% 
  rename(cell_status_at_frame = filamented_at_frame) %>% 
  fill(initial_time, end_time, .direction = "up") %>% 
  filter(time >= initial_time) %>% 
  mutate(
    cell_status_at_frame = as.character(cell_status_at_frame),
    cell_status_at_frame = replace_na(cell_status_at_frame, "Dead"),
    cell_status_at_frame = factor(cell_status_at_frame, levels = c("Not filamented", "Filamented", "Dead")),
    time = factor(time)
  ) %>% 
  identity()
```

#### With dead

```{r}
status_time_df %>% 
  ggplot(aes(x = time, fill = cell_status_at_frame)) +
  geom_bar(position = "fill", stat = "count", width = 1) +
  geom_vline(xintercept = c("60", "140"), linetype = "dashed") +
  facet_grid(experiment_id ~ .) +
  scale_x_discrete(expand = c(0, 0)) +
  scale_y_continuous(expand = c(0, 0), labels = scales::percent) +
  theme(
    panel.spacing.y = unit(1, "lines")
  ) +
  geom_text(
    data = data.frame(
      x = c(7.5, 15.5),
      y = c(0.75, 0.25),
      label = c("Start", "End"),
      experiment_id = "Plasmid"
    ),
    mapping = aes(x = x, y = y, label = label),
    size = 6,
    hjust = 0L,
    vjust = 0L,
    colour = "white",
    inherit.aes = FALSE
  ) +
  labs(
    x = "Time (minutes)",
    y = "Percentage of cells",
    fill = "Cell status"
  ) +
  guides(
    color = guide_legend(ncol = 2),
    fill = guide_legend(ncol = 2)
  ) +
  NULL
```

#### Without dead

```{r}
status_time_df %>% 
  filter(cell_status_at_frame != "Dead") %>% 
  ggplot(aes(x = time, fill = cell_status_at_frame)) +
  geom_bar(position = "fill", stat = "count", width = 1) +
  geom_vline(xintercept = c("60", "140"), linetype = "dashed") +
  facet_grid(experiment_id ~ .) +
  scale_x_discrete(expand = c(0, 0)) +
  scale_y_continuous(expand = c(0, 0), labels = scales::percent) +
  theme(
    panel.spacing.y = unit(1, "lines")
  ) +
  geom_text(
    data = data.frame(
      x = c(7.5, 15.5),
      y = c(0.75, 0.75),
      label = c("Start", "End"),
      experiment_id = "Plasmid"
    ),
    mapping = aes(x = x, y = y, label = label),
    size = 6,
    hjust = 0L,
    vjust = 0L,
    colour = "white",
    inherit.aes = FALSE
  ) +
  labs(
    x = "Time (minutes)",
    y = "Percentage of cells",
    fill = "Cell status"
  ) +
  scale_fill_hue(direction = -1, h.start = 90) +
  #scale_fill_manual(values = c("#619CFF", "#00BA38")) +
  NULL
```

### Differences between groups

### Metrics distributions

```{r metric_charts, warning=FALSE, message=FALSE}
processed_lineages_df_2 %>% 
  select(experiment_id, cell_status, time, length, gfp, ds_red) %>% 
  pivot_longer(
    cols = c(length, gfp, ds_red),
    names_to = "metric"
  ) %>%
  mutate(
    metric = case_when(
      metric == "ds_red" ~ "DsRed",
      metric == "gfp" ~ "GFP",
      metric == "length" ~ "Length"
    )
  ) %>% 
  group_by(experiment_id, cell_status, time, metric) %>%
  summarise(
    ci = list(mean_cl_normal(value)),
    .groups = "drop"
  ) %>% 
  unnest(cols = c(ci)) %>% 
  ggplot(aes(x = time, y = y, ymin= ymin, ymax=ymax, color = cell_status)) +
  annotate("rect", xmin=60, xmax=140, ymin=-Inf, ymax=Inf, alpha=1/2, color = "transparent", fill = "#FCB565") +
  geom_smooth(method = "loess") +
  facet_grid(metric ~ experiment_id, scales = "free_y") +
  labs(
    x = "Time (minutes)",
    y = "Value",
    color = "Cell status"
  ) +
  theme_bw() +
  theme(legend.position = "top") +
  guides(
    color = guide_legend(ncol = 2),
    fill = guide_legend(ncol = 2)
  ) +
  scale_color_hue(direction = -1, h.start = 90) +
  scale_fill_hue(direction = -1, h.start = 90)
```

### GFP Survival probability

#### Calculation

```{r survival_probability}
gfp_control_hist <- processed_lineages_df_2 %>% 
  ggplot(aes(x = gfp)) +
  geom_histogram(bins = 100) +
  facet_grid(experiment_id ~ .)
  
gfp_hist_data <- gfp_control_hist %>% 
  ggplot_build() %>% 
  pluck("data", 1) %>% 
  select(PANEL, count, x, xmin, xmax) %>% 
  mutate(experiment_id = case_when(
    PANEL == 1 ~ "Chromosome",
    PANEL == 2 ~ "Plasmid"
  )) %>% 
  as_tibble()

gfp_breaks <- gfp_hist_data %>% 
  group_by(experiment_id) %>% 
  summarise(
    breaks = unique(c(.$xmin, last(.$xmax))),
    .groups = "drop"
  )

gfp_survival_probability_df <- processed_lineages_df_2 %>%
  group_by(experiment_id, id, filamented_track) %>% 
  summarise(
    initial_gfp = first(gfp),
    is_long_track = first(centered_frame) < unique(centered_antibiotic_start_frame) &&
      last(centered_frame) > unique(centered_antibiotic_end_frame),
    .groups = "drop"
  ) %>% 
  filter(is_long_track) %>%
  group_by(experiment_id, filamented_track) %>% 
  group_modify(~{
    
    current_data <- .x
    current_experiment <- current_data %>%
      pull(experiment_id) %>% 
      unique()
    
    current_control_gfp <- gfp_hist_data %>% 
      filter(experiment_id == current_experiment) %>% 
      pull(count)
    
    current_breaks <- gfp_breaks %>% 
      filter(experiment_id == current_experiment) %>% 
      pull(breaks)
    
    tibble(
      plot = list(
        ggplot(data = current_data, aes(x = initial_gfp)) +
        geom_histogram(
          breaks = current_breaks
        )
      )
    ) %>%
      mutate(
        counts = map(plot, ggplot_build),
        counts = map(counts, pluck, "data", 1),
        counts = map(counts, add_column, control_count = current_control_gfp),
        counts = map(counts, select, gfp = x, control_count, count)
      ) %>% 
      select(-plot)
  }, .keep = TRUE) %>%
  unnest(counts) %>% 
  mutate(
    survival_probability = count / control_count,
    #survival_probability = survival_probability / max(survival_probability, na.rm = TRUE)
  ) %>%
  #filter(survival_probability != 0) %>%
  glimpse() %>%
  identity()
```

#### Points

```{r survival_probability_plot}
gfp_survival_probability_df %>% 
  ggplot(aes(x = gfp, y = survival_probability, color = filamented_track)) +
  geom_point() +
  geom_rug(alpha = 1 / 2) +
  scale_y_continuous(labels = scales::percent) +
  #facet_grid(. ~ experiment_id) +
  theme_bw() +
  theme(
    legend.position = "top"
  ) +
  labs(
    x = "Initial GFP (log10)",
    y = "Survival probability",
    color = "Cell status",
    linetype = "Experiment"
  ) +
  scale_color_hue(direction = -1, h.start = 90) +
  scale_fill_hue(direction = -1, h.start = 90) +
  facet_grid(. ~ experiment_id) +
  coord_cartesian(ylim = c(c(0, 0.04))) +
  NULL
```

### Length Survival probability

#### Calculation

```{r}
length_control_hist <- processed_lineages_df_2 %>% 
  ggplot(aes(x = length)) +
  geom_histogram(bins = 100) +
  facet_grid(experiment_id ~ .)
  
length_hist_data <- length_control_hist %>% 
  ggplot_build() %>% 
  pluck("data", 1) %>% 
  select(PANEL, count, x, xmin, xmax) %>% 
  mutate(experiment_id = case_when(
    PANEL == 1 ~ "Chromosome",
    PANEL == 2 ~ "Plasmid"
  )) %>% 
  as_tibble()

length_breaks <- length_hist_data %>% 
  group_by(experiment_id) %>% 
  summarise(
    breaks = unique(c(.$xmin, last(.$xmax))),
    .groups = "drop"
  )

length_survival_probability_df <- processed_lineages_df_2 %>%
  group_by(experiment_id, id, filamented_track) %>% 
  summarise(
    initial_length = first(length),
    is_long_track = first(centered_frame) < unique(centered_antibiotic_start_frame) &&
      last(centered_frame) > unique(centered_antibiotic_end_frame),
    .groups = "drop"
  ) %>% 
  filter(is_long_track) %>%
  group_by(experiment_id, filamented_track) %>% 
  group_modify(~{
    
    current_data <- .x
    current_experiment <- current_data %>%
      pull(experiment_id) %>% 
      unique()
    
    current_control_length <- length_hist_data %>% 
      filter(experiment_id == current_experiment) %>% 
      pull(count)
    
    current_breaks <- length_breaks %>% 
      filter(experiment_id == current_experiment) %>% 
      pull(breaks)
    
    tibble(
      plot = list(
        ggplot(data = current_data, aes(x = initial_length)) +
        geom_histogram(
          breaks = current_breaks
        )
      )
    ) %>%
      mutate(
        counts = map(plot, ggplot_build),
        counts = map(counts, pluck, "data", 1),
        counts = map(counts, add_column, control_count = current_control_length),
        counts = map(counts, select, length = x, control_count, count)
      ) %>% 
      select(-plot)
  }, .keep = TRUE) %>%
  unnest(counts) %>% 
  mutate(
    survival_probability = count / control_count,
    #survival_probability = survival_probability / max(survival_probability, na.rm = TRUE)
  ) %>%
  #filter(survival_probability != 0) %>%
  glimpse() %>%
  identity()
```


#### Points

```{r length_survival_probability_plot}
length_survival_probability_df %>% 
  filter(survival_probability != 1) %>% 
  ggplot(aes(x = length, y = survival_probability, color = filamented_track)) +
  geom_point() +
  geom_rug(alpha = 1 / 2) +
  scale_y_continuous(labels = scales::percent) +
  facet_grid(. ~ experiment_id) +
  theme_bw() +
  theme(
    legend.position = "top"
  ) +
  labs(
    x = "Initial length",
    y = "Survival probability",
    color = "Cell status",
    linetype = "Experiment"
  ) +
  coord_cartesian(ylim = c(c(0, 0.03))) +
  scale_color_hue(direction = -1, h.start = 90) +
  scale_fill_hue(direction = -1, h.start = 90)
```
