==> Rmarkdown/2.0-jvelezmagic-tidymodel.Rmd <==
---
title: "Experiment Data Exploration"
author: "Jesús Vélez Santiago"
date: "`r format(Sys.Date(), '%Y-%m')`"
output: 
  html_document:
    theme: readable
    highlight: kate
    toc: true
    toc_float: true
    toc_depth: 3
    code_folding: show
    self_contained: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = TRUE,
  message = FALSE,
  fig.align = "center",
  # dev = "svg",
  fig.retina = 2
)
```

## Libraries

```{r libraries, message=FALSE}
library(tidyverse)
library(tidymodels)
library(themis)
library(here)
library(vip)
library(glue)
```

## Load data

```{r load_data}
lineages_files <- here("data", "processed", "lineages.tsv")
lineages_df <- read_tsv(lineages_files, show_col_types = FALSE) %>%
  glimpse()
```

## Minimal preprocessing

```{r minimal_preprocessing}
processed_lineages_df <- lineages_df %>%
  mutate(
    gfp = log10(gfp),
    ds_red = log10(ds_red),
    across(contains("filamented_"), ~ factor(
      .x,
      levels = c(FALSE, TRUE),
      labels = c("Not filamented", "Filamented")
    ))
  ) %>%
  add_count(experiment_id, trap_id, track_id, time) %>%
  filter(n == 1) %>%
  select(-n) %>%
  glimpse()
```

## Cell status

```{r}
cell_status_data <- processed_lineages_df %>%
  group_by(experiment_id, trap_id, track_id, filamented_track) %>%
  summarize(
    # Identify inflection points.
    tmp_first_true = which.max(filamented_at_frame), # SOS state.
    # Extract GFP
    initial_gfp = first(gfp),
    sos_gfp = gfp[tmp_first_true],
    end_gfp = last(gfp),
    # Extract length
    initial_length = first(length),
    sos_length = length[tmp_first_true],
    end_length = last(length),
    # Extract time points
    initial_time = first(centered_frame),
    sos_time = centered_frame[tmp_first_true],
    end_time = last(centered_frame),
    # Calculate utilities.
    tmp_life_time = end_time - initial_time,
    tmp_start_experiment = first(centered_antibiotic_start_frame),
    tmp_end_experiment = first(centered_antibiotic_end_frame),
    tmp_exists_before_or_during_experiment = initial_time < first(centered_antibiotic_end_frame),
    tmp_exists_after_experiment = end_time > first(centered_antibiotic_end_frame),
    survive = tmp_exists_before_or_during_experiment && tmp_exists_after_experiment,
    survive = ifelse(survive, "Alive", "Dead"),
    survive = factor(survive),
    .groups = "drop"
  ) %>%
  filter(
    # Remove missing cells before the start of experiment.
    end_time > tmp_start_experiment,
    # Remove new born cell after antibiotic exposure.
    initial_time < tmp_end_experiment
  ) %>%
  mutate(
    #filamented_track = as.character(filamented_track),
    status = interaction(filamented_track, survive),
    status = as.character(status),
    status = factor(status)
  ) %>%
  mutate(
    tmp_invalid_sos = filamented_track == "Not filamented" & initial_time == sos_time,
    sos_gfp = case_when(
      tmp_invalid_sos ~ NaN,
      TRUE ~ sos_gfp
    ),
    sos_length = case_when(
      tmp_invalid_sos ~ NaN,
      TRUE ~ sos_length
    ),
    sos_time = case_when(
      tmp_invalid_sos ~ NaN,
      TRUE ~ sos_time
    )
  ) %>%
  select(-contains("tmp"), -trap_id, -track_id) %>%
  identity() %>%
  glimpse()
```

## Default for plots

```{r}
theme_set(
  theme_bw() +
  theme(
    legend.position = "top"
  )
)
```

## What is going to be classified?

```{r}
cell_status_data %>%
  count(experiment_id, status) %>% 
  mutate(status = str_replace(status, "\\.", " ")) %>% 
  ggplot(aes(x = experiment_id, y = n, fill = status)) +
  geom_bar(position = "fill", stat = "identity", color = "white") +
  scale_y_continuous(labels = scales::percent) +
  guides(fill = guide_legend(nrow = 2)) +
  labs(
    x = "Experiment",
    y = "Percentage of cells", 
    fill = "Cell status"
  ) +
  scale_fill_hue(direction = -1, h.start = 90)
```



## Simple EDA

```{r}
cell_status_data %>%
  mutate(life_time = (end_time - initial_time) * 10) %>%
  count(experiment_id, filamented_track, life_time) %>%
  ggplot(aes(x = as.factor(life_time), y = n, fill = filamented_track)) +
  geom_bar(position = "fill", stat = "identity", width = 1) +
  scale_x_discrete(expand = c(0, 0)) +
  scale_y_continuous(expand = c(0, 0), labels = scales::percent) +
  facet_grid(experiment_id ~ .) +
  theme_bw() +
  theme(
    legend.position = "top",
    panel.spacing.y = unit(1, "lines")
  ) +
  labs(
    x = "Cell life time",
    y = "Percentage of cells",
    fill = "Cell status"
  ) +
  scale_fill_hue(direction = -1, h.start = 90)
```

```{r}
cell_status_data %>%
  mutate(life_time = (end_time - initial_time) * 10) %>%
  group_by(experiment_id, filamented_track, life_time, survive) %>%
  summarize(
    n = n(),
    initial_length = median(initial_length),
    sos_length = median(sos_length),
    end_length = median(end_length),
    .groups = "drop"
  ) %>%
  pivot_longer(
    cols = contains("length"),
    names_to = "length_type"
  ) %>%
  mutate(
    length_type = factor(length_type, levels = c("initial_length", "sos_length", "end_length"), labels = c("Initial", "SOS", "End")),
  ) %>%
  drop_na() %>%
  ggplot(aes(x = life_time, y = value)) +
  geom_line(aes(group = life_time)) +
  geom_point(aes(color = length_type, shape = survive), alpha = 1 / 1) +
  facet_grid(filamented_track ~ experiment_id) +
  theme_bw() +
  theme(
    legend.position = "top"
  ) +
  labs(
    x = "Cell life time",
    y = "Length value",
    color = "Length type",
    shape = "Cell status"
  ) +
  scale_color_hue(direction = -1, h.start = 90)
```

```{r}
cell_status_data %>%
  mutate(life_time = (end_time - initial_time) * 10) %>%
  group_by(experiment_id, filamented_track, life_time, survive) %>%
  summarize(
    n = n(),
    initial_gfp = median(initial_gfp),
    sos_gfp = median(sos_gfp),
    end_gfp = median(end_gfp),
    .groups = "drop"
  ) %>%
  pivot_longer(
    cols = contains("gfp"),
    names_to = "gfp_type"
  ) %>%
  mutate(
    gfp_type = factor(gfp_type, levels = c("initial_gfp", "sos_gfp", "end_gfp"), labels = c("Initial", "SOS", "End"))
  ) %>%
  drop_na() %>%
  ggplot(aes(x = life_time, y = value)) +
  geom_line(aes(group = life_time)) +
  geom_point(aes(color = gfp_type, shape = survive), alpha = 1 / 1) +
  facet_grid(filamented_track ~ experiment_id) +
  theme_bw() +
  theme(
    legend.position = "top"
  ) +
  labs(
    x = "Cell life time",
    y = "GFP value",
    color = "GFP type",
    shape = "Cell status"
  ) +
  scale_fill_hue(direction = -1, h.start = 90)
```

## All data

### Build a model

```{r}
cells_df <- cell_status_data %>%
  select(experiment_id, status, contains(c("gfp", "length")), -contains("sos")) %>%
  glimpse()
```

```{r}
set.seed(42)
cells_boot <- bootstraps(cells_df, times = 10)

cells_boot
```

```{r}
cells_rec <- recipe(status ~ ., data = cells_df) %>%
  update_role(experiment_id, new_role = "Id") %>%
  step_dummy(all_nominal(), -all_outcomes(), -has_role("Id")) %>%
  step_zv(all_predictors()) %>%
  step_normalize(all_predictors()) %>%
  step_smote(status)

cells_rec
```

```{r}
cells_prep <- prep(cells_rec)
juice(cells_prep)
```

```{r}
rf_spec <- rand_forest(trees = 100) %>%
  set_engine("ranger") %>%
  set_mode("classification")
```

```{r}
cells_wf <- workflow() %>%
  add_recipe(cells_rec) %>%
  add_model(rf_spec)

cells_wf
```

```{r}
cells_res <- fit_resamples(
  cells_wf,
  resamples = cells_boot,
  control = control_resamples(save_pred = TRUE)
)
```

### Explore results

```{r}
cells_res %>%
  collect_metrics()
```

```{r}
cells_res %>%
  collect_predictions() %>%
  conf_mat(truth = status, .pred_class)
```

```{r}
cells_res %>%
  collect_predictions() %>%
  group_by(id) %>%
  ppv(status, .pred_class)
```

```{r}
rf_fit <- rf_spec %>%
  set_engine("ranger", importance = "permutation") %>%
  fit(
    status ~ .,
    data = juice(cells_prep) %>%
      select(-experiment_id)
  )

rf_fit %>%
  vip(geom = "point")
```

```{r}
cells_pred <- cells_res %>%
  collect_predictions() %>%
  mutate(correct = status == .pred_class) %>%
  left_join(
    cells_df %>%
      mutate(.row = row_number())
  )

cells_pred
```

```{r}
cells_pred %>%
  separate(status, c("is_filamented", "status"), sep = "\\.") %>%
  mutate(
    is_filamented = factor(is_filamented, levels = c("Not filamented", "Filamented")),
    status = factor(status, levels = c("Dead", "Alive"))
  ) %>%
  ggplot(aes(x = end_length, y = initial_gfp)) +
  stat_summary_hex(
    aes(z = as.integer(correct)),
    fun = "mean",
    alpha = 0.9,
    bins = 20
  ) +
  facet_grid(experiment_id ~ is_filamented) +
  scale_fill_gradient(labels = scales::percent) +
  labs(
    x = "End length",
    y = "Initial GFP",
    fill = "Percent classified correctly"
  )
```

```{r}
cells_pred %>%
  separate(status, c("is_filamented", "status"), sep = "\\.") %>%
  mutate(
    is_filamented = factor(is_filamented, levels = c("Not filamented", "Filamented")),
    status = factor(status, levels = c("Dead", "Alive"))
  ) %>%
  #pull(is_filamented)
  ggplot(aes(x = end_length, y = initial_gfp)) +
  geom_hex(aes(fill = status), alpha = 1 / 3, bins = 20) +
  facet_grid(experiment_id ~ is_filamented) +
  labs(
    x = "End length",
    y = "Initial GFP",
    fill = "Cell status"
  )
```

## Just filamented plasmid cells

```{r}
filamented_cells_data <- cell_status_data %>%
  filter(experiment_id == "Plasmid", filamented_track == "Filamented") %>%
  glimpse()
```

### Build a model

```{r}
cells_df <- filamented_cells_data %>%
  mutate(diff_sos_initial = sos_time - initial_time) %>%
  select(experiment_id, survive, diff_sos_initial, contains(c("gfp", "length"))) %>%
  glimpse()
```

```{r}
set.seed(42)
cells_boot <- bootstraps(cells_df, times = 10)

cells_boot
```

```{r}
cells_rec <- recipe(survive ~ ., data = cells_df) %>%
  update_role(experiment_id, new_role = "Id") %>%
  step_dummy(all_nominal(), -all_outcomes(), -has_role("Id")) %>%
  step_zv(all_predictors()) %>%
  step_normalize(all_predictors()) %>%
  step_smote(survive)

cells_rec
```

```{r}
cells_prep <- prep(cells_rec)
juice(cells_prep)
```

```{r}
rf_spec <- rand_forest(trees = 100) %>%
  set_engine("ranger") %>%
  set_mode("classification")
```

```{r}
cells_wf <- workflow() %>%
  add_recipe(cells_rec) %>%
  add_model(rf_spec)

cells_wf
```

```{r}
cells_res <- fit_resamples(
  cells_wf,
  resamples = cells_boot,
  control = control_resamples(save_pred = TRUE)
)
```

### Explore results

```{r}
cells_res %>%
  collect_metrics()
```

```{r}
cells_res %>%
  collect_predictions() %>%
  conf_mat(truth = survive, .pred_class)
```

```{r}
cells_res %>%
  collect_predictions() %>%
  group_by(id) %>%
  ppv(survive, .pred_class)
```

```{r}
rf_fit <- rf_spec %>%
  set_engine("ranger", importance = "permutation") %>%
  fit(
    survive ~ .,
    data = juice(cells_prep) %>%
      select(-experiment_id)
  )

rf_fit %>%
  vip(geom = "point")
```

```{r}
cells_pred <- cells_res %>%
  collect_predictions() %>%
  mutate(correct = survive == .pred_class) %>%
  left_join(
    cells_df %>%
      mutate(.row = row_number())
  ) %>%
  mutate(
    survive = factor(survive, levels = c("Dead", "Alive"))
  )

cells_pred
```

```{r}
cells_pred %>%
  ggplot(aes(x = initial_length, y = end_length)) +
  stat_summary_hex(
    aes(z = as.integer(correct)),
    fun = "mean",
    alpha = 0.9,
    bins = 20
  ) +
  facet_grid(experiment_id ~ .) +
  scale_fill_gradient(labels = scales::percent) +
  labs(
    x = "Initial length",
    y = "End length",
    fill = "Percent classified correctly"
  )
```

```{r}
cells_pred %>%
  ggplot(aes(x = initial_length, y = end_length)) +
  geom_hex(aes(fill = survive), alpha = 1 / 3, bins = 20) +
  facet_grid(experiment_id ~ .) +
  labs(
    x = "Initial length",
    y = "End length",
    fill = "Cell status"
  )
```

## Just filamented chromosome cells

```{r}
filamented_cells_data <- cell_status_data %>%
  filter(filamented_track == "Filamented", experiment_id == "Chromosome") %>%
  glimpse()
```

### Build a model

```{r}
cells_df <- filamented_cells_data %>%
  mutate(diff_sos_initial = sos_time - initial_time) %>%
  select(experiment_id, diff_sos_initial, survive, contains(c("gfp", "length"))) %>%
  glimpse()
```

```{r}
set.seed(42)
cells_boot <- bootstraps(cells_df, times = 10)

cells_boot
```

```{r}
cells_rec <- recipe(survive ~ ., data = cells_df) %>%
  update_role(experiment_id, new_role = "Id") %>%
  step_dummy(all_nominal(), -all_outcomes(), -has_role("Id")) %>%
  step_zv(all_predictors()) %>%
  step_normalize(all_predictors()) %>%
  step_smote(survive)

cells_rec
```

```{r}
cells_prep <- prep(cells_rec)
juice(cells_prep)
```

```{r}
rf_spec <- rand_forest(trees = 100) %>%
  set_engine("ranger") %>%
  set_mode("classification")
```

```{r}
cells_wf <- workflow() %>%
  add_recipe(cells_rec) %>%
  add_model(rf_spec)

cells_wf
```

```{r}
cells_res <- fit_resamples(
  cells_wf,
  resamples = cells_boot,
  control = control_resamples(save_pred = TRUE)
)
```

### Explore results

```{r}
cells_res %>%
  collect_metrics()
```

```{r}
cells_res %>%
  collect_predictions() %>%
  conf_mat(truth = survive, .pred_class)
```

```{r}
cells_res %>%
  collect_predictions() %>%
  group_by(id) %>%
  ppv(survive, .pred_class)
```

```{r}
rf_fit <- rf_spec %>%
  set_engine("ranger", importance = "permutation") %>%
  fit(
    survive ~ .,
    data = juice(cells_prep) %>%
      select(-experiment_id)
  )

rf_fit %>%
  vip(geom = "point")
```

```{r}
cells_pred <- cells_res %>%
  collect_predictions() %>%
  mutate(correct = survive == .pred_class) %>%
  left_join(
    cells_df %>%
      mutate(.row = row_number())
  ) %>%
  mutate(
    survive = factor(survive, levels = c("Dead", "Alive"))
  )

cells_pred
```

```{r}
cells_pred %>%
  ggplot(aes(x = end_gfp, y = end_length)) +
  stat_summary_hex(
    aes(z = as.integer(correct)),
    fun = "mean",
    alpha = 0.9,
    bins = 20
  ) +
  facet_grid(experiment_id ~ .) +
  scale_fill_gradient(labels = scales::percent) +
  labs(
    x = "End GFP",
    y = "End length",
    fill = "Percent classified correctly"
  )
```

```{r}
cells_pred %>%
  ggplot(aes(x = end_gfp, y = end_length)) +
  geom_hex(aes(fill = survive), alpha = 1 / 3, bins = 20) +
  facet_grid(experiment_id ~ .) +
  labs(
    x = "End gfp",
    y = "End length",
    fill = "Cell status"
  )
```

## Time

```{r}
processed_lineages_df %>% 
  mutate(division = ifelse(division, TRUE, FALSE)) %>% 
  select(experiment_id, trap_id, track_id, length, centered_frame, centered_antibiotic_start_frame, division, filamented_track, filamented_at_frame) %>% 
  filter(centered_frame <= centered_antibiotic_start_frame, division) %>% 
  group_by(experiment_id, trap_id, track_id, filamented_track) %>% 
  summarise(
    last_division_frame = max(centered_frame * division),
    time_since_last_division = centered_antibiotic_start_frame - last_division_frame,
    across(.cols = c(centered_frame, length, division), last),
    .groups = "drop"
  ) %>% 
  glimpse() %>% 
  count(experiment_id, filamented_track, time_since_last_division) %>% 
  ggplot(aes (x = factor(time_since_last_division * 10), y = n, fill = filamented_track)) +
  geom_bar(position = "fill", stat = "identity", color = "white")+
  facet_wrap(experiment_id ~ .) +
  scale_y_continuous(labels = scales::percent) +
  labs(
    x = "Time since last division before or equal to the start of the experiment (minutes)",
    y = "Percentage of cells",
    fill = "Cell status"
  ) +
  scale_fill_hue(direction = -1, h.start = 90) +
  NULL
```

```{r}
processed_lineages_df %>% 
  mutate(division = ifelse(division, TRUE, FALSE)) %>% 
  select(experiment_id, trap_id, track_id, length, centered_frame, centered_antibiotic_start_frame, division, filamented_track, filamented_at_frame) %>% 
  mutate(
    filamented_at_frame = case_when(
      filamented_at_frame == "Filamented" ~ TRUE,
      TRUE ~ FALSE
    )
  ) %>% 
  filter(centered_frame <= centered_antibiotic_start_frame) %>% 
  group_by(experiment_id, trap_id, track_id, filamented_track) %>% 
  summarise(
    last_division_frame = max((centered_frame + 1) * division) - 1,
    frames_since_last_division = centered_antibiotic_start_frame - last_division_frame,
    divided = !all(division == FALSE),
    not_previous_filamented = !any(filamented_at_frame),
    across(.cols = c(centered_frame, filamented_at_frame, length, division), list),
    .groups = "drop"
  ) %>%
  #unnest() %>% 
  filter(divided, not_previous_filamented) %>% 
  count(experiment_id, filamented_track, frames_since_last_division) %>% 
  ggplot(aes (x = factor(frames_since_last_division * 10), y = n, fill = filamented_track)) +
  geom_bar(position = "fill", stat = "identity", color = "white")+
  facet_wrap(experiment_id ~ .) +
  scale_y_continuous(labels = scales::percent) +
  labs(
    x = "Time since last division before or equal to the start of the experiment (minutes)",
    y = "Percentage of cells",
    fill = "Cell status"
  ) +
  scale_fill_hue(direction = -1, h.start = 90) +
  NULL
```

```{r}
survived_cells_df <- processed_lineages_df %>% 
  group_by(experiment_id, trap_id, track_id, filamented_track) %>% 
  summarise(
    exists_before_experiment = first(centered_frame) < first(centered_antibiotic_start_frame),
    exists_before_or_during_experiment = first(centered_frame) <= first(centered_antibiotic_end_frame),
    exists_after_experiment = last(centered_frame) > first(centered_antibiotic_end_frame),
    survived_soft = exists_before_or_during_experiment && exists_after_experiment,
    survived_strict = exists_before_experiment && exists_after_experiment,
    .groups = "drop"
  ) %>% 
  glimpse() %>% 
  identity()
```


```{r}
processed_lineages_df %>% 
  left_join(survived_cells_df, by = c("experiment_id", "trap_id", "track_id", "filamented_track")) %>%
  mutate(division = ifelse(division, TRUE, FALSE)) %>% 
  filter(centered_frame <= centered_antibiotic_start_frame) %>% 
  group_by(experiment_id, trap_id, track_id, filamented_track, survived_strict) %>% 
  summarize(
    # Status
    divided = !all(division == FALSE),
    not_previous_filamented = !any(filamented_at_frame == "Filamented"),
    # Time
    last_division_frame = max((centered_frame + 1) * division) - 1,
    frames_since_last_division_to_experiment = first(centered_antibiotic_start_frame) - last_division_frame,
    .groups = "drop"
  ) %>% 
  filter(divided, not_previous_filamented) %>% 
  mutate(
    time_since_last_division_to_experiment = factor(frames_since_last_division_to_experiment * 10),
    survived_strict = factor(survived_strict, labels = c("Not survived", "Survived"))
  ) %>% 
  count(experiment_id, filamented_track, survived_strict, time_since_last_division_to_experiment) %>% 
  identity() %>% 
  ggplot(aes(x = time_since_last_division_to_experiment, y = n, fill = filamented_track)) +
  geom_bar(position = "dodge", stat = "identity", color = "white") +
  facet_grid(experiment_id ~ survived_strict, scales = "free_y") +
  labs(
    x = "Minutes difference between the start of the antibiotic entry and the last division",
    y = "Number of cells",
    fill = "Cell status"
  ) +
  scale_fill_hue(direction = -1, h.start = 90) +
  NULL
```

```{r}
processed_lineages_df %>% 
  left_join(survived_cells_df, by = c("experiment_id", "trap_id", "track_id", "filamented_track")) %>%
  mutate(division = ifelse(division, TRUE, FALSE)) %>% 
  filter(centered_frame <= centered_antibiotic_start_frame) %>% 
  group_by(experiment_id, trap_id, track_id, filamented_track, survived_strict) %>% 
  summarize(
    # Status
    divided = !all(division == FALSE),
    not_previous_filamented = !any(filamented_at_frame == "Filamented"),
    # Time
    last_division_frame = max((centered_frame + 1) * division) - 1,
    frames_since_last_division_to_experiment = first(centered_antibiotic_start_frame) - last_division_frame,
    .groups = "drop"
  ) %>% 
  filter(divided, not_previous_filamented) %>% 
  mutate(
    time_since_last_division_to_experiment = factor(frames_since_last_division_to_experiment * 10),
    survived_strict = factor(survived_strict, labels = c("Not survived", "Survived"))
  ) %>% 
  count(experiment_id, filamented_track, survived_strict, time_since_last_division_to_experiment) %>% 
  identity() %>% 
  ggplot(aes(x = time_since_last_division_to_experiment, y = n, fill = filamented_track)) +
  geom_bar(position = "fill", stat = "identity", color = "white") +
  facet_grid(experiment_id ~ survived_strict, scales = "free_y") +
  scale_y_continuous(labels = scales::percent) +
  labs(
    x = "Minutes difference between the start of the antibiotic entry and the last division",
    y = "Percentage of cell",
    fill = "Cell status"
  ) +
  scale_fill_hue(direction = -1, h.start = 90) +
  NULL
```

```{r}
processed_lineages_df %>% 
  left_join(survived_cells_df, by = c("experiment_id", "trap_id", "track_id", "filamented_track")) %>%
  mutate(division = ifelse(division, TRUE, FALSE)) %>% 
  filter(centered_frame <= centered_antibiotic_start_frame) %>% 
  group_by(experiment_id, trap_id, track_id, filamented_track, survived_strict) %>% 
  summarize(
    # Status
    divided = !all(division == FALSE),
    not_previous_filamented = !any(filamented_at_frame == "Filamented"),
    # Time
    last_division_frame = max((centered_frame + 1) * division) - 1,
    frames_since_last_division_to_experiment = first(centered_antibiotic_start_frame) - last_division_frame,
    .groups = "drop"
  ) %>% 
  filter(divided, not_previous_filamented) %>% 
  mutate(
    time_since_last_division_to_experiment = factor(frames_since_last_division_to_experiment * 10),
    survived_strict = factor(survived_strict, labels = c("Not survived", "Survived"))
  ) %>% 
  count(experiment_id, filamented_track, survived_strict, time_since_last_division_to_experiment) %>% 
  identity() %>% 
  ggplot(aes(x = time_since_last_division_to_experiment, y = n, fill = survived_strict)) +
  geom_bar(position = "fill", stat = "identity", color = "white") +
  facet_grid(experiment_id ~ filamented_track, scales = "free_y") +
  scale_y_continuous(labels = scales::percent) +
  labs(
    x = "Minutes difference between the start of the antibiotic entry and the last division",
    y = "Percentage of cell",
    fill = "Cell status"
  ) +
  scale_fill_hue(direction = -1, h.start = 90) +
  NULL
```
